% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/memalpha.R
\name{character_data}
\alias{character_data}
\title{Information about characters in Star Trek}
\usage{
character_data(chname)
}
\arguments{
\item{chname}{a character's name}
}
\value{
a \code{data.frame} giving a character's gender, affiliation, species and rank
}
\description{
Handily get all the information on a single character from memory-alpha.org
}
\examples{
# get information for Major Kira
character_data("Kira_Nerys")
# get information for Worf
character_data("Worf")
}
\author{
A Andrew M MacDonald
}

