% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{dev_std_crtlk}
\alias{dev_std_crtlk}
\title{For TCGA data get human pathway data and creates a measure of standard deviations among pathways}
\usage{
dev_std_crtlk(dataFilt, pathway)
}
\arguments{
\item{dataFilt}{TCGA matrix}

\item{pathway}{pathway data}
}
\value{
a matrix value for each pathway
}
\description{
st_dv creates a matrix with standard deviation for pathways
}
\examples{
\dontrun{
path<-getKEGGdata(KEGG_path="Transcript")
cancer <- "TCGA-BRCA"
PlatformCancer <- "Illumina HiSeq"
tumour<-c("TCGA-BH-A0DL-01A-11R-A115-07","TCGA-AO-A03P-01A-11R-A00Z-07")
normal<-c("TCGA-BH-A209-11A-42R-A157-07","TCGA-E9-A1N4-11A-33R-A14M-07") 
TCGA_matrix<-get_TCGAdata(cancer,PlatformCancer,tumour,normal,
                                       patha = "exampleData")
cross_talk_st_dv<-dev_std_crtlk(dataFilt=TCGA_matrix,pathway=path)}
}

