% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds9}
\alias{ds9}
\title{Transcripts for Star Trek: Deep Space Nine}
\format{An object of class \code{list} of length 173.}
\usage{
ds9
}
\description{
Transcripts for Star Trek: Deep Space Nine
}
\keyword{datasets}
