% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/memalpha.R
\name{episodes}
\alias{episodes}
\title{Information about episodes in a season (or seasons) of Star Trek}
\usage{
episodes(show = c("TOS", "TNG", "DS9", "VOY", "ENT"), seasons = "all")
}
\arguments{
\item{show}{a valid identifier for a show}

\item{seasons}{which season (or seasons) should be included. Overflow will cover all seasons of a show, with no error. Default to \code{"all"}, which shows all seasons.}
}
\value{
a (list of) \code{data.frame}(s) of episodes, season per list item/\code{data.frame}, one episode per row.
}
\description{
Handily get all the episodes in a season (plus extra info) or all the episodes in a given season of Star Trek.
}
\examples{
# get episodes for season 5 of Deep Space 9
episodes("DS9",5)
# get episodes for seasons 2 and 3 of the Original Series
episodes("TOS", 2:3)
}
\author{
David L. Miller
}

