% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{GE_matrix}
\alias{GE_matrix}
\title{Get human KEGG pathway data and a gene expression matrix in order to obtain a matrix with the gene expression for only pathways given in input .}
\usage{
GE_matrix(DataMatrix, pathway)
}
\arguments{
\item{DataMatrix}{gene expression matrix (eg.TCGA data)}

\item{pathway}{pathway data as provided by getKEGGdata}
}
\value{
a matrix for each pathway ( gene expression level belong to that pathway)
}
\description{
GE_matrix creates a matrix of gene expression for pathways given by the user.
}
\examples{
list_path_plot<-GE_matrix(DataMatrix=tumo[,1:2],pathway=path)
}

