% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{get_TCGAdata}
\alias{get_TCGAdata}
\title{Get TCGA data.}
\usage{
get_TCGAdata(cancer, PlatformCancer, tumour, normal, patha)
}
\arguments{
\item{cancer}{cancer type, See TCGAbiolinks package}

\item{PlatformCancer}{platform, type See TCGAbiolinks package}

\item{tumour}{barcode samples with label for example tumour}

\item{normal}{barcode samples with label for example normal}

\item{patha}{directory name}
}
\value{
dataframe with gene-gene (or protein-protein interactions)
}
\description{
get_TCGAdata creates a data frame with network data. 
Network category can be filtered among: physical interactions, co-localization, genetic interactions and shared protein domain.
}
\examples{
cancer <- "TCGA-BRCA"
PlatformCancer <- "Illumina HiSeq"
tumour<-c("TCGA-BH-A0DL-01A-11R-A115-07","TCGA-AO-A03P-01A-11R-A00Z-07")
normal<-c("TCGA-BH-A209-11A-42R-A157-07","TCGA-E9-A1N4-11A-33R-A14M-07") 
TCGA_matrix<-get_TCGAdata(cancer,PlatformCancer,tumour,normal,
                                       patha = "exampleData")
}

