% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getKEGGdata}
\alias{getKEGGdata}
\title{Get human KEGG pathway data.}
\usage{
getKEGGdata(KEGG_path)
}
\arguments{
\item{KEGG_path}{variable}
}
\value{
dataframe with human pathway data
}
\description{
getKEGGdata creates a data frame with human KEGG pathway. Columns are the pathways and rows the genes inside those pathway
}
\examples{
path<-getKEGGdata(KEGG_path="Transcript")
}

