% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getdata.R
\name{getNETdata}
\alias{getNETdata}
\title{Get network data.}
\usage{
getNETdata(network, organism = NULL)
}
\arguments{
\item{network}{variable. The user can use the following parameters 
based on the network types to be used. PHint for Physical_interactions,
COloc for Co-localization, GENint for Genetic_interactions and
SHpd for Shared_protein_domains}

\item{organism}{organism==NULL default value is homo sapiens}
}
\value{
dataframe with gene-gene (or protein-protein interactions)
}
\description{
getNETdata creates a data frame with network data. 
Network category can be filtered among: physical interactions, co-localization, genetic interactions and shared protein domain.
}
\examples{
organism="Saccharomyces_cerevisiae"
netw<-getNETdata(network="SHpd",organism)
}

