% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{hamm_dist}
\alias{hamm_dist}
\title{Hamming distance}
\usage{
hamm_dist(list_pat)
}
\arguments{
\item{list_path}{pathway data}
}
\value{
a symmetric matrix value with a distance among pathways
}
\description{
hamm_dist creates a matrix with Hamming distance values for pathways
}
\examples{
path<-getKEGGdata(KEGG_path="Transcript")
netw<-getNETdata(network="SHpd")
list_path<-list_path_net(net_type=netw,pathway=path)
hamm_distance<-hamm_dist(list_path)
}

