% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{list_path_net}
\alias{list_path_net}
\title{Get human KEGG pathway data and network data in order to define the common gene.}
\usage{
list_path_net(net_type, pathway)
}
\arguments{
\item{net_type}{network data as provided by getNETdata}

\item{pathway}{pathway data as provided by getKEGGdata}
}
\value{
a list of genes for each pathway (interacting genes belong to that pathway)
}
\description{
list_path_net creates a list of interacting genes for each human pathway.
}
\examples{
list_path<-list_path_net(net_type=netw,pathway=path)
}

