% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarTrek.r
\docType{data}
\name{list_path_plot}
\alias{list_path_plot}
\title{A matrix of gene expression for pathways given by the user.}
\format{A data frame with rows and variables}
\description{
A matrix of gene expression for pathways given by the user.
}
\keyword{internal}

