% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{path_net}
\alias{path_net}
\title{Get human KEGG pathway data and network data in order to define the common gene.}
\usage{
path_net(pathway, net_type)
}
\arguments{
\item{pathway}{pathway data as provided by getKEGGdata}

\item{net_type}{network data as provided by getNETdata}
}
\value{
a list of network data for each pathway (interacting genes belong to that pathway)
}
\description{
path_net creates a list of network data for each human pathway. The network data will be generated when interacting genes belong to that pathway.
}
\examples{
lista_net<-path_net(pathway=path,net_type=netw)
}

