% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{plotting_cross_talk}
\alias{plotting_cross_talk}
\title{Get human KEGG pathway data and a gene expression matrix we obtain a matrix with the gene expression for only pathways given in input .}
\usage{
plotting_cross_talk(DataMatrix, pathway, path_matrix)
}
\arguments{
\item{DataMatrix}{gene expression matrix (eg.TCGA data)}

\item{pathway}{pathway data as provided by getKEGGdata}

\item{path_matrix}{output of the function GE_matrix}
}
\value{
a plot for pathway cross talk
}
\description{
plotting_matrix creates a matrix of gene expression for pathways given by the user.
}
\examples{
mt<-plotting_cross_talk(DataMatrix=tumo[,1:2],pathway=path,path_matrix=list_path_plot)
}

