% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{process_matrix}
\alias{process_matrix}
\title{Process matrix TCGA data after the selection of pairwise pathway}
\usage{
process_matrix(measure, list_perf)
}
\arguments{
\item{measure}{matrix with measure of cross-talk among pathways}

\item{list_perf}{output of the function select_class}
}
\value{
a gene expression matrix for case study 1
}
\description{
processing gene expression matrix
}

