% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarTrek.r
\docType{data}
\name{score_euc_dist}
\alias{score_euc_dist}
\title{Score Matrix of pairwise pathway using euclidean distance}
\format{A data frame with rows and variables}
\description{
Score Matrix of pairwise pathway using euclidean distance
}
\keyword{internal}

