% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{SelectedSample}
\alias{SelectedSample}
\title{Select the class of TCGA data}
\usage{
SelectedSample(Dataset, typesample)
}
\arguments{
\item{Dataset}{gene expression matrix}

\item{typesample}{the labels of the samples (e.g. tumor,normal)}
}
\value{
a gene expression matrix of the samples with specified label
}
\description{
select two labels from ID barcode
}
\examples{
tumo<-SelectedSample(Dataset=Data_CANCER_normUQ_filt,typesample="tumor")[,2]
}

