% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StarTrek.r
\docType{package}
\name{StarTrek}
\alias{StarTrek}
\alias{StarTrek-package}
\title{Download data}
\description{
StarTrek allows you to Download data of samples from StarTrek
}
\details{
The functions you're likely to need from \pkg{StarTrek} is
\code{path_star}
Otherwise refer to the vignettes to see
how to format the documentation.
}

