% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{svm_classification}
\alias{svm_classification}
\title{SVM classification for each feature}
\usage{
svm_classification(TCGA_matrix, tumour, normal, nfs)
}
\arguments{
\item{TCGA_matrix}{gene expression matrix}

\item{tumour}{barcode samples for a class}

\item{normal}{barcode samples for another class}

\item{nfs}{nfs split data into a training  and test set}
}
\value{
a symmetric matrix value with a distance among pathways
}
\description{
svm class creates a list with auc value
}
\examples{
tumo<-SelectedSample(Dataset=BRCA_rnaseqv2,typesample="tumor")[,1:30]
norm<-SelectedSample(Dataset=BRCA_rnaseqv2,typesample="normal")[,1:30]
score_euc_dist<-euc_dist_crtlk(dataFilt=BRCA_rnaseqv2,path)
nf <- 60
res_class<-svm_classification(TCGA_matrix=BRCA_rnaseqv2,nfs=nf,normal=colnames(norm),tumour=colnames(tumo))
}

