% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{tau_dist}
\alias{tau_dist}
\title{For TCGA data get human pathway data and creates a measure of cross-talk among pathways}
\usage{
tau_dist(dataFilt, pathway)
}
\arguments{
\item{dataFilt}{TCGA matrix}

\item{pathway}{pathway data}
}
\value{
a matrix value for each pathway
}
\description{
tau_dist creates a matrix with tau distance for pairwise pathways
}
\examples{
score_tau_dist<-tau_dist(dataFilt=Data_CANCER_normUQ_filt,path)
}

