% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tng}
\alias{tng}
\title{Transcripts for Star Trek: The Next Generation}
\format{An object of class \code{list} of length 176.}
\usage{
tng
}
\description{
Transcripts for Star Trek: The Next Generation
}
\keyword{datasets}
