% Generated by roxygen2 (4.0.2.9000): do not edit by hand
% Please edit documentation in R/warp_speed.R
\name{warp_speed}
\alias{warp_speed}
\title{Warp Speed}
\usage{
warp_speed(warpfactor, originalseries = FALSE)
}
\arguments{
\item{warpfactor}{Warp factor to be converted}

\item{originalseries}{Defaults to FALSE. Warp speed is different in The
Original Series than in later series. Set to TRUE if you want to use the
format for TOS.}
}
\description{
Converts warp factor to speed in meters/second
}
\author{
Kara Woo
}
\references{
\url{http://www.anycalculator.com/warpcalculator.htm}
}

