% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{dev_std_crtlk}
\alias{dev_std_crtlk}
\title{For TCGA data get human pathway data and creates a measure of standard deviations among pathways}
\usage{
dev_std_crtlk(dataFilt, pathway)
}
\arguments{
\item{dataFilt}{TCGA matrix}

\item{pathway}{pathway data}
}
\value{
a matrix value for each pathway
}
\description{
st_dv creates a matrix with standard deviation for pathways
}
\examples{
cross_talk_st_dv<-dev_std_crtlk(dataFilt=BRCA_rnaseqv2,pathway=path)
}

