% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{euc_dist_crtlk}
\alias{euc_dist_crtlk}
\title{For TCGA data get human pathway data and creates a measure of cross-talk among pathways}
\usage{
euc_dist_crtlk(dataFilt, pathway)
}
\arguments{
\item{dataFilt}{TCGA matrix}

\item{pathway}{pathway data}
}
\value{
a matrix value for each pathway
}
\description{
euc_dist_crtlk creates a matrix with euclidean distance for pairwise pathways
}
\examples{
score_euc_dista<-euc_dist_crtlk(dataFilt=tumo[,1:2],path)
}

