% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{proc_path}
\alias{proc_path}
\title{Get human KEGG pathway data.}
\usage{
proc_path(mer)
}
\arguments{
\item{mer}{output for example of select_path_carb}
}
\value{
dataframe with human pathway data
}
\description{
getKEGGdata creates a data frame with human KEGG pathway. Columns are the pathways and rows the genes inside those pathway
}

