% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{score_median}
\alias{score_median}
\title{For TCGA data get human pathway data and creates a matrix with the median of genes for each pathway.}
\usage{
score_median(dataFilt, pathway)
}
\arguments{
\item{dataFilt}{TCGA matrix}

\item{pathway}{pathway data}
}
\value{
a matrix value for each pathway
}
\description{
score_median creates a matrix with a summarized value for each pathway
}
\examples{
med<-score_median(dataFilt=Data_CANCER_normUQ_filt,path)
}

