% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{select_class}
\alias{select_class}
\title{Select the class of TCGA data}
\usage{
select_class(auc.df, cutoff)
}
\arguments{
\item{auc.df}{list of AUC value}

\item{cutoff}{cut-off for AUC value}
}
\value{
a gene expression matrix with only pairwise pathway with a particular cut-off
}
\description{
select two labels from ID barcode
}

