% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{st_dv}
\alias{st_dv}
\title{For TCGA data get human pathway data and creates a measure of standard deviations among pathways}
\usage{
st_dv(DataMatrix, pathway)
}
\arguments{
\item{DataMatrix}{TCGA matrix}

\item{pathway}{pathway data}
}
\value{
a matrix value for each pathway
}
\description{
st_dv creates a matrix with standard deviation for pathways
}
\examples{
stand_dev<-st_dv(DataMatrix=tumo[,1:2],pathway=path)
}

