% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/path_star.R
\name{svm_classification}
\alias{svm_classification}
\title{SVM classification for each feature}
\usage{
svm_classification(Data_CANCER_mRNA_tumor_HighStage, nfs)
}
\arguments{
\item{Data_CANCER_mRNA_tumor_HighStage}{Data_CANCER_mRNA_tumor_HighStage}

\item{nfs}{nfs split data into a training  and test set}
}
\value{
a symmetric matrix value with a distance among pathways
}
\description{
svm class creates a list with auc value
}
\examples{
\dontrun{
Data_me=read.table('C:/Users/UserInLab05/Desktop/GSE22216_series_matrix.txt',header = TRUE)
nf <- 60
res_class<-svm_classification(Data_me,nf)
}
}

